We have to set the ContentLength explicitly
Since Int.MaxValue is ~2.147 GB in Scala, and individual S3 objects can
be up to 5 TB, we could get a fetch entry which is larger than this
value allows.

The Content-Length that comes from getObject() is a Long, not an Int,
to support the larger files -- that's what we should pass around.