Improve the BagMatcher message for dangling fetch entries
Saying "it's not in the bag" is a bit ambiguous (and confused at least
one person) -- if it's referred to in the fetch file, we expect it not
to be in the bag!  So clarify that it's not in the bag manifest.

Also add a test for multiple dangling entries on the same path.