Always have at least 3 retries on every queue
If we deploy a new version of a service, ECS stops any instances of the
old service.  This means we lose any work that was in-flight.

If the SQS queue has retries >1, the message will eventually be retried
(albeit after a delay, from the queue visibility timeout).  If we don't
have retries, the message will get dropped and not retried, and the
ingest will be eternally processing.