Add support for downloading a bag directly to a compressed file
In Archivematica, we've discovered that the cost of preparing a bag
for redownload is disproportionately at the compression step.
Downloading a 600MB takes 2–3 minutes, compressing it takes 10–15.

We suspect this is the cost of making lots of little reads back to the
filesystem, so we're going to try compressing as we stream them out
of the storage service, which seems like it should be more efficient.