Rename "archive bucket" to "Ireland replica"
This reflects a naming change elsewhere; we now talk about the
Ireland replica instead of the archive bucket (and -archive no longer
appears in the bucket names).

Additionally, it turns out the bag register was reading from
the Ireland replica!  This is bad behaviour -- let's read from the
primary replica instead.