Rename the "ingests drop" bucket to "unpacked bags"
This better reflects what we're using the bucket for, and disambiguates
the "drop bucket" name for use by services (e.g. Goobi, Archivematica)
that upload a package to a bucket, which triggers some more work.
The bag where we temporarily store unpacked bags isn't that!

Additionally, the definition of the bucket moves into the stack module,
as this only contains transient data, nothing permanent we want to keep.