Refer to the storage-manifests VHS as "vhs_manifests" consistently
It had a bunch of inconsistent names, including:

* manifest
* manifests
* archive_manifests
* vhs_archive_manifest

More consistency should make this code easier to read and follow, and
reduce ambiguity.  Also less meaningless use of the word "archive"!

This has no effect on the actual infra, just the Terraform state file.

This commit also applies a consistent ordering to these variables/outputs,
for ease of comparison across files.