Add a blanket rule that disallows slashes in space/externalIdentifier
Supporting slashes adds a bunch of edge cases and weirdnesses that
we'd need to check (e.g. around S3 paths and URL path components).
If we had time to test it thoroughly, we could do it, but for now reject
any ingest that tries to use "a/b" or "b/c" as a space or external ID.