Make ExternalIdentifier and StorageSpace regular classes
This disables the automatic case class derivation.  Why is this important?
Because we already have data in our databases/stores that uses the
formatting

    ExternalIdentifier("123") -> json: "123"

and the automatic derivation would do

    ExternalIdentifier("123") -> json: {"underlying": "123"}

We used to get this provided by AnyVal, but because we've added runtime
assertions about the values, we can't use AnyVal any more.