Use .coercedXmap instead of .iso for DynamoFormat
Using DynamoFormat.iso[A, B] is only appropriate if every instance of A
can be an instance of B, and vice versa.  That's no longer the case for
these classes: every external identifier can be a string, but the empty
string isn't a valid ID.

This caused the tests to fail whenever you tried to serialise an
ExternalIdentifier using DynamoFormat, at the following line:

    def iso[A, B](r: B => A)(w: A => B)(implicit f: DynamoFormat[B]) = new DynamoFormat[A] {
      ...
      final override val default = f.default.map(r)
    }

The default DynamoFormat[String] is the empty string, and trying to map
that to ExternalIdentifier would throw.