Add more restrictions on external identifier and storage space
Since we're now doing validation of these strings, cover a couple of
edge cases that would cause annoyance if you put them in the storage
service.  In particular:

*   Don't allow slashes in a place that would cause double slashes in
    the S3 key (because that causes weirdness in the S3 Console, which
    is how a lot of people will browse the storage)
*   Don't allow an identifier to end with `/versions` (because that
    conflicts with the bags API endpoint)
*   Don't allow an identifier to include `/v1`, `/v2`, `/v3`, ...
    (because that makes the S3 bucket harder to navigate)

The storage service itself will cope fine with any of these things, but
they'll cause confusion for the humans who have to use it!

If we find we actually need identifiers that look like this, we can
consider removing them, and accept the possible confusion that might
result.  But since we don't expect to use them any time soon, this
should prevent programming bugs elsewhere causing confusion in the
storage service!