Add tests for the unpacker message "unpacked X bytes from Y files"
The logic here is getting non-trivial, and it would be good to have some
more robust tests.  Also, adding a test exposed a bug!  If you had a bag
with a single file*, it would print

    Unpacked 4 MB from 1 file()

The extra () is because we didn't define an `else` branch on the `if`
statement, so a Unit fell out, and the string for Unit is ().  So now
we print the correct message:

    Unpacked 4 MB from 1 file

This is also laying the groundwork for a further improvement to these
messages, which will come in the next commit.

*Although that would be unusual; most bags would at minimum have four
files -- the manifest, tag manifest, bagit.txt and bag-info.txt.