Format the file count in unpacker messages with a comma
Before:  Unpacked 7 GB from 49684 files
After:   Unpacked 7 GB from 49,684 files

The comma makes it a little easier for humans to parse and read.  This
number is only indicative -- if a machine wants the file count, it can
inspect the storage manifest at the end of the pipeline -- so this just
makes these messages a bit friendlier.