Attempt to fix the flaky bag verifier test, maybe?
Sometimes the first test in this file fails, complaining that there's
one more file in the manifest than in the Payload-Oxum.  Checking, I saw
the following unexpected line:

    badDigest  badName

in the manifest of the bag crated for the test that *should* be working.
I've no idea how it's leaking, but I've put a note here and changed the
name to emphasise that this is a bad builder, don't use this elsewhere.