Allow getting the latest BagVersion from the storage manifest dao
This is a more efficient lookup than calling listVersions() or
getLatest(), because it only has to look in the DynamoDB table, and it
doesn't serialise the full storage manifest out of S3.

This is preliminary work for a change to the bags API to allow us to
get cached responses from an S3 bucket, rather than fetching a complete
response every time.