Tweak the internal routing logic in the bags API
If we know the user didn't supply a ?version= or a ?before= parameter,
skip trying to parse the version and go straight to the response.

This lets us clean up some odd type signatures (Try[Option[T]]) and
ditch a use of Option.get -- now the type system knows that Option is
filled.