Check the S3 cache before returning a response
If we have a large bag to return, the request will often time out before
we've responded – the bags API returns a 500 error, the user doesn't
get their bag.

This patch modifies the bags API to look for a cached response first,
and gets a pre-signed URL from S3 if so, skipping the VHS.

This means the first request for a bag might fail, but hopefully it will
populate the cache and eventually start working.