Fix the bug causing OOM errors in the S3StreamReader
Because we were making a ranged GET request, S3 would report the
content length of every file as 1.  That would cause the reader to read
straight to the end, even if it was an enormous thing, and buffer the
entire object in memory.

Using .abort() and not a ranged GET avoids this issue *and* means we
don't get the warning about "not all bytes read from the stream".