Introduce the notion of "ManifestFile" and an associated parser
This commit lays the groundwork for parsing all the manifest files in
a bag (not just the SHA256 manifest), with a side-effect of reducing the
amount of duplicated information we pass around.

If you look at the previous BagManifest definition:

    case class BagManifest(
      checksumAlgorithm: HashingAlgorithm,
      files: Seq[BagFile]
    )

the information about the checksum algorithm is stored twice: once on
the manifest itself, once on the BagFile.  This seems odd, given that
the algorithm is a property of the entire manifest file.  We put it on
the individual BagFile instances to aid in verification, but that's not
really correct.

This commit introduces a notion of a ManifestFile and a VerifiableChecksum.
The checksum can store values for all four checksum algorithms, and the
ManifestFile is a map from path to checksum, which better represents
the semantics of the manifest file:

    sealed trait CombinedManifestFile {
      val files: Map[BagPath, VerifiableChecksum]
    }

We're no longer carrying around lots of redundant information on the
individual entries in the manifest.

Then we can separate the payload and tag manifests at the type system
level, which have the same format but are semantically different.

This commit introduces a new parser for the generic manifest format,
which can then combine manifests from different algorithms (e.g. MD5 and
SHA1).

Additionally, it improves the level of checking we do on manifests:

*   The checksum should be a hex-encoded string.  If it's not, we can
    reject the manifest file immediately, and the regex that inspects
    the line now filters for hex strings.

*   The BagIt spec says that a manifest should refer to each file
    exactly once, but we never enforced this previously.  We do now.

*   The BagIt spec says that each manifest should refer to the same set
    of files, but we never enforced this previously.  This is now
    possible.