Start tracking a Map[BagPath, FetchMetadata] in BagFetch
Once we know entries in a fetch.txt are unique, we can record them as
a map rather than a list of entries.  This opens the door to using the
data in a cleaner way in other parts of the code, because we can compare
the keys (paths) with a list of paths elsewhere.

In particular, we can ask "does fetch.txt contain path P?" in a more
efficient way than looping over the entire fetch file.