Only hold the checksum algorithm on the Manifest, not the Entries
Previously the BagManifest would hold the checksum algorithm (MD5, SHA1, etc.)
on the top level *and* inside all the BagFile/Checksum entries, for each
line in the manifest.

The algorithm is a file-level property, not a property of the individual
entries, so remove it from the entries and get it from the file when
needed -- don't duplicate the information.

As a nice bonus, this means we can delete tests that check we handle
inconsistency correctly -- it's no longer for possible for that
inconsistency to occur.