Fix a flaky test in the bag verifier feature test
This was an absolute pain to figure out -- a test that fails sometimes,
but with no apparent error or reason.

When I finally sat down to look at it more carefully, I found some
suspicious clues:

-   I saw the bag starting to verify multiple times, and multiple
    "Verification started" ingest events being sent.
-   The bag seemed to complete verification successfully; there were no
    errors complaining about a bad checksum or anything.
-   The comment about payloads being received multiple times should have
    been more of a hint.
-   The exact assertion that was failing was around the size of the DLQ,
    not the queue.

So here's the issue:

The size of the bag is randomly chosen at test time, based on a couple of
factors:

-   How many payload files are in the bag (5 to 50)
-   The path to each of those files (1 to 5 folders deep)
-   How much data is in each file (1 to 256 characters)

If the bag was sufficiently big, it would miss the SQS visibility
timeout and be resent, and eventually the test would fail.  You could
make the test fail reliably by increasing the lower bound (or fixing)
the size of the bag to be suitably large.

Increasing the timeout on the SQS queue makes the problem go away.