Try to make this bag register test less flaky
This test failed in Travis for non-obvious reasons:
https://travis-ci.org/github/wellcomecollection/storage-service/jobs/674771412

It looks like SQS delivered the message twice, which could be a bug or
could be because the bag register failed to complete within the visibility
timeout.  Bumping it is easy, so let's try that and hope the bug doesn't
happen again.