Reduce the boilerplate in BagVerifierTest
A lot of this test code is repeated: set up the bucket, create the bag,
get the verification result, see it's an instance of IngestFailed[_].
The "interesting" assertions are harder to see.  Pushing it into a couple
of helper methods makes the tests easier to follow.

More practically, we need to modify how the verifier is called for
https://github.com/wellcomecollection/platform/issues/4513; specifically
to pass the bucket and storage space.  Having the test boilerplate in
a helper means less diff churn elsewhere.