Split the verification of file size from verification of checksum
This is another bit of refactoring in prep for the storage tiering work.
When an object gets lifecycled to Glacier, we can't get the length by
inspecting the input stream; we'll have to do a HeadObject call.  Splitting
this into a separate function now will make it easier to change later.