Check we close all the InputStream's that get opened
We need to do this to avoid exhausting the verifier's resources, but
previously there was no test that we were doing so.  I deliberately
deleted this line in the previous commit to see if the tests caught it,
and they didn't.

While I'm here, let's reduce the level of branching in this function --
one level is plenty, no need for two.