Simplify the creation of StreamStore in the unpacker tests
We used to have a distinction between two types of wrapped InputStream:
InputStreamWithLength and InputStreamWithLengthAndMetadata.

The unpacker only cares about length, not metadata.  The pre-built stores
from the storage library only came in metadata variants, and we didn't
have any sort of covariance set up – a StreamStore[ISWithLength] and
StreamStore[ISWithLengthAndMetadata] are completely distinct things.
That means we had to wrap a StreamStore[Metadata] store to get
a StreamStore[Length].

We no longer have both types of InputStream, just WithLength, so we can
remove all this wrapper code.