Create this instance of TagManifestFileFinder lazily
If you don't make this value lazy, you get an error when you start the
app via Main:

    Exception in thread "main" scala.UninitializedFieldError:
    Uninitialized field: bag_register/services/s3/S3StorageManifestService.scala: 28

I haven't fully debugged this error and I'm not sure why it occurs here
but not in the tests; I do know that this change fixes it.  This is
breaking the pipeline and doesn't add any tech debt, so I'm happy to
leave it without digging further.