Split the tests: one for trailing slash, one without
Using the trailing slash is the more correct way to do this sort of
list prefixing -- but it means we don't catch the issue in the old
(broken) code.  Having a test that runs it both ways catches the bug.

This is a copy/paste of the existing test with just a tiny variation.

Co-authored-by: Alice Fuzier Cayla <alicefuzier@gmail.com>