Draw a line under existing, known failures
I have reviewed all the failures that happened in the past and decided
they're not worth looking into further -- but if we get a new unexplained
failure, we'll want to dig in further.

Adding a hardcoded list of known, uninteresting failures is an easy way
to have new failures be surfaced but old failures ignored.