Reduce the repetition in the unpacker error handling
Because the S3 SDK is wrapping exceptions from the underlying Java I/O
libraries for reading compressed streams, it can hand off error handling
to the base Unpacker trait.  This keeps the error handling in one place
and means we have less places where we repeat the user-facing error
messages.