Add some more context for this AzureFixityChecker comment
When I wrote that comment, I didn't realise the NAT instance might have
been at fault.

The current value of the constant is working for us and I don't want to
break a working system, but it's useful context for the next person who
works on this code/wants to change this constant.