Remove the Alpakka dependencies
We're getting an error in the storage service in prod:

    java.lang.ClassNotFoundException:
    com.amazonaws.metrics.internal.cloudwatch.DefaultMetricCollectorFactory

Stack Overflow suggests this may be caused by loading multiple versions
of the AWS SDK.  Using sbt-dependency-graph (https://github.com/sbt/sbt-dependency-graph)
reveals that Alpakka-S3 is causing us to do just that:

    +-uk.ac.wellcome:common_2.12:0.1.0-SNAPSHOT [S]
      +-com.lightbend.akka:akka-stream-alpakka-s3_2.12:1.1.2 [S]
      | +-com.amazonaws:aws-java-sdk-core:1.11.476 (evicted by: 1.11.504)
      | +-com.amazonaws:aws-java-sdk-core:1.11.504

which made me wonder: do we need that dependency at all?

Signs point to no, at least not directly, so let's bin it.