Remove the default metrics from the replifier
We did this for our other services in https://github.com/wellcomecollection/storage-service/pull/450,
and now we're getting a lot of noise in the logs because we took the
Metrics SDK out of the classpath.  We don't need these metrics, so remove
them.  Hopefully this will make the logs less noisy and make it easier
to debug real issues.