Make the Azure replicator optional
For the DAMS prototype project, we don't want to require an Azure account
and do replication into Azure.  Make it optional, so if the caller doesn't
specify an Azure Blob container, we won't replicate there.

I had to do a small state change in our stacks to make this a no-op plan:

    $ terraform state mv module.stack_prod.module.replicator_verifier_azure module.stack_prod.module.replicator_verifier_azure[0]

I haven't tested this with the "no Azure" configuration, because I don't
want to affect our existing environments -- I'm hoping to try it in the
DAMS prototype account soon.