Improve verification of the Bag Declaration (bagit.txt)
We defer to Scala to split lines for us (`String.lines`), so we don't
have to worry about supporting different line endings.  This means we get
support for Windows line endings for "free".

This approach also means we can provide slightly better error messages
when we don't accept the Bag Declaration as supplied.