Order API Gateway resources before API Deployments
API Gateway resources (such as aws_api_gateway_method,
aws_api_gateway_resource, etc) are now explicitly ordered to be applied
before the aws_api_gateway_deployment which (indirectly) references
them. In addition, the Deployment is re-created when any of these
resources are modified.

Together, this ordering and re-creation of the Deployment ensures that
all resources are visible from a deployment after a terraform apply,
both after an initial and subsequent applies.

Without the state of API Gateway resources being reflected in a
aws_api_gateway_deployment's triggers, changes to API Gateway resources
won't be visible to API users, because a deployment must be re-created
to make changes to resources live. Similarly, without ordering, a
Deployment can be created before the resources it should contain exit.