Explicitly manage the AWS API Gateway Stage
The aws_api_gateway_deployment docs recommend against having it auto-
manage a Stage using the stage_name argument. Leaving stage_name
unspecified achieves this.

Previously, the config was using the stage_name argument on
aws_api_gateway_deployment and explicitly creating a Stage. In a new
deploy, the plan would show that no v1 stage existed, so the plan would
create the explicit v1 Stage, and the aws_api_gateway_deployment would
create a v1 stage as a result of its stage_name argument. After the
first run, the v1 stage would exist, so the explicit Stage would not
be re-created, so no conflict would occur.

Creating an explicit stage resource should be error-free so long as the
stage_name argument for aws_api_gateway_deployment is not used.

Warning: when updating an existing deployment, the existing "v1"
Stage will need to be manually deleted the first time only, otherwise
the apply will fail with the previously-encountered error:

	Error creating API Gateway Stage: ConflictException: Stage already exists

This is because aws_api_gateway_deployment.v1 needs to use
create_before_destroy. As a result, when deploying the explicit Stage
added in this commit, a new Deployment to replace
aws_api_gateway_deployment.v1 is created before the old Deployment is
deleted, which results in two "v1" Stages existing concurrently, which
isn't permitted.