Give the replicators List* permissions
The replicator is a service that does bulk copying of objects from one
S3 bucket to another.  But we don't want to overwrite/replace objects in
the destination bucket if they already exist

* if they're the same we can save ourselves a write
* if they're different then something has gone extremely wrong

So the replicator queries for the destination object before it starts the
copy.

What happens if there's no existing object?  If you call GetObject on a
non-existent object, one of two things can happen:

*   If you have List* permissions, S3 returns a 404 Not Found
*   If you don't have List* permissions, S3 returns a 401 Not Authorized
    (to avoid leaking information about bucket contents in 404s)

It turns out we'd given the replicator Get* and Put* permissions on the
destination bucket, but not List*, so all these requests for non-existent
objects were 401'ing.  Adding the right permissions means they'll 404,
which is more correct.

Note: we do have some code to reduce the number of unnecessary requests
here, but it's a fairly conservative check (better to check than not),
and the cost of a few extra requests hasn't been unwieldy so far.